pyproject.toml
[tool.poetry]
name = "best-clips-youtube"
version = "0.1.0"
description = "best clips exxtraction"
authors = ["ashrithjacob <ashrithjacob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
supabase = "^2.7.4"
python-dotenv = "^1.0.1"
certifi = "^2024.8.30"
openai = "^1.48.0"
torch = "2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
